chore: 기타 코드 수정
