+++ src/components/elisa/3d-models/Cheering1RuModel.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 import { Html, useGLTF } from "@react-three/drei";
 import { GLTF } from "three-stdlib";
 import { IClickedMeshInfo } from "./StadiumModel";
-import MeshLabel from "./MeshLabel";
+import MeshLabel from "../MeshLabel";
 
 type NodeKeys =
   | "Mesh5748_1Ru-cheering_zone-309"
