+++ src/components/elisa/3d-models/SkeletonModel.tsx
