+++ src/components/elisa/3d-models/Sky1RuModel.tsx
@@ -3,18 +3,21 @@ import { useEffect, useState } from "react";
 import { Html, useGLTF } from "@react-three/drei";
 import { GLTF } from "three-stdlib";
 import { IClickedMeshInfo } from "./StadiumModel";
-import MeshLabel from "./MeshLabel";
+import MeshLabel from "../MeshLabel";
 
 type NodeKeys =
-  | "Mesh23780_Y-box_zone-114"
-  | "Mesh23940_Y-box_zone-115"
-  | "Mesh24131_Y-box_zone-214"
-  | "Mesh24329_Y-box_zone-311"
-  | "Mesh24550_Y-box_zone-312"
-  | "Mesh24648_Y-box_zone-215"
-  | "Mesh26049_Y-box_zone-310"
-  | "Mesh26234_Y-box_zone-213"
-  | "Mesh26341_Y-box_zone-113";
+  | "Mesh10445_Sky-1ru_zone-401"
+  | "Mesh10445_Sky-1ru_zone-402"
+  | "Mesh10445_Sky-1ru_zone-403"
+  | "Mesh10445_Sky-1ru_zone-405"
+  | "Mesh10445_Sky-1ru_zone-406"
+  | "Mesh10445_Sky-1ru_zone-407"
+  | "Mesh10445_Sky-1ru_zone-408"
+  | "Mesh15721_Sky-1ru_zone-410"
+  | "Mesh9812_Sky-1ru_zone-404"
+  | "Mesh10445_Sky-1ru_zone-409"
+  | "Mesh10445_Sky-1ru_zone-411"
+  | "Mesh10445_Sky-1ru_zone-412";
 
 type GLTFResult = GLTF & {
   nodes: {
@@ -29,34 +32,64 @@ type MeshData = {
 };
 
 const meshesData: MeshData[] = [
-  { name: "Mesh23780_Y-box_zone-114", position: [-633.48, 56.168, -297.805] },
-  { name: "Mesh23940_Y-box_zone-115", position: [-609.484, 56.167, -395.034] },
-  { name: "Mesh24131_Y-box_zone-214", position: [-766.962, 90.389, -332.551] },
-  { name: "Mesh24329_Y-box_zone-311", position: [-905.723, 132.763, -367.435] },
-  { name: "Mesh24550_Y-box_zone-312", position: [-882.458, 132.764, -461.133] },
-  { name: "Mesh24648_Y-box_zone-215", position: [-743.575, 90.377, -426.586] },
-  { name: "Mesh26049_Y-box_zone-310", position: [-928.716, 133.153, -276.234] },
-  { name: "Mesh26234_Y-box_zone-213", position: [-789.662, 91.12, -249.489] },
-  { name: "Mesh26341_Y-box_zone-113", position: [-655.978, 57.012, -218.884] },
+  { name: "Mesh10445_Sky-1ru_zone-401", position: [-737.017, 239.18, 560.506] },
+  {
+    name: "Mesh10445_Sky-1ru_zone-402",
+    position: [-743.393, 239.566, 488.915],
+  },
+  {
+    name: "Mesh10445_Sky-1ru_zone-403",
+    position: [-752.707, 239.976, 380.518],
+  },
+  { name: "Mesh10445_Sky-1ru_zone-405", position: [-775.214, 239.96, 91.657] },
+  {
+    name: "Mesh10445_Sky-1ru_zone-406",
+    position: [-785.097, 239.224, -54.314],
+  },
+  {
+    name: "Mesh10445_Sky-1ru_zone-407",
+    position: [-784.296, 237.213, -189.634],
+  },
+  {
+    name: "Mesh10445_Sky-1ru_zone-408",
+    position: [-763.859, 240.732, -313.978],
+  },
+  {
+    name: "Mesh15721_Sky-1ru_zone-410",
+    position: [-714.377, 252.987, -570.564],
+  },
+  { name: "Mesh9812_Sky-1ru_zone-404", position: [-763.976, 239.949, 235.664] },
+  {
+    name: "Mesh10445_Sky-1ru_zone-409",
+    position: [-727.512, 239.49, -437.893],
+  },
+  {
+    name: "Mesh10445_Sky-1ru_zone-411",
+    position: [-595.412, 204.771, -643.678],
+  },
+  {
+    name: "Mesh10445_Sky-1ru_zone-412",
+    position: [-539.837, 204.736, -713.59],
+  },
 ];
 
-export function YBoxModel({
+export function Sky1RuModel({
   hides,
   areaName,
   showModal,
   handleMeshHover,
   handleMeshClick,
 }: any) {
-  const { nodes, materials } = useGLTF("/models/y-box.glb") as GLTFResult;
+  const { nodes, materials } = useGLTF("/models/sky-1ru.glb") as GLTFResult;
 
   const [isHovered, setIsHovered] = useState(false);
   const [hoveredMesh, setHoveredMesh] = useState<IClickedMeshInfo | null>(null);
   const [clickedMesh, setClickedMesh] = useState<IClickedMeshInfo | null>(null);
 
-  const defaultColor = nodes["Mesh23780_Y-box_zone-114"]
+  const defaultColor = nodes["Mesh10445_Sky-1ru_zone-401"]
     .material as THREE.MeshStandardMaterial;
   const hoverColor = defaultColor.clone();
-  hoverColor.color.set("#EC9341");
+  hoverColor.color.set("#211C79");
 
   const onMeshClick = (info: IClickedMeshInfo): void => {
     handleMeshClick(info);
@@ -124,4 +157,4 @@ export function YBoxModel({
   );
 }
 
-useGLTF.preload("/models/y-box.glb");
+useGLTF.preload("/models/sky-1ru.glb");
