+++ src/components/elisa/3d-models/StadiumSkeleton.tsx
@@ -1,0 +1,7 @@
+// import { ReactComponent as MyIcon } from '/images/stadium.svg';
+
+// const StadiumSkeleton = () => {
+//   return <div>StadiumSkeleton</div>;
+// };
+
+// export default StadiumSkeleton;
