+++ src/components/elisa/3d-models/TvingTableModel.tsx
@@ -1,11 +1,11 @@
 import * as THREE from "three";
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { Html, useGLTF } from "@react-three/drei";
 import { GLTF } from "three-stdlib";
 import { IClickedMeshInfo } from "./StadiumModel";
-import MeshLabel from "./MeshLabel";
+import MeshLabel from "../MeshLabel";
 
-type NodeKeys = "Mesh6965_1Ru-exiting";
+type NodeKeys = "Mesh2887_Tving-table";
 
 type GLTFResult = GLTF & {
   nodes: {
@@ -14,23 +14,23 @@ type GLTFResult = GLTF & {
   materials: { [key: string]: THREE.MeshStandardMaterial };
 };
 
-export function Exiting1RuModel({
+export function TvingTableModel({
   hides,
   areaName,
   showModal,
   handleMeshHover,
   handleMeshClick,
 }: any) {
-  const { nodes, materials } = useGLTF("/models/exiting-1ru.glb") as GLTFResult;
+  const { nodes, materials } = useGLTF("/models/tving-table.glb") as GLTFResult;
 
   const [isHovered, setIsHovered] = useState(false);
   const [hoveredMesh, setHoveredMesh] = useState<IClickedMeshInfo | null>(null);
   const [clickedMesh, setClickedMesh] = useState<IClickedMeshInfo | null>(null);
 
-  const defaultColor = nodes["Mesh6965_1Ru-exiting"]
+  const defaultColor = nodes["Mesh2887_Tving-table"]
     .material as THREE.MeshStandardMaterial;
   const hoverColor = defaultColor.clone();
-  hoverColor.color.set("#36A09A");
+  hoverColor.color.set("#DB7390");
 
   const onMeshClick = (info: IClickedMeshInfo): void => {
     handleMeshClick(info);
@@ -57,7 +57,7 @@ export function Exiting1RuModel({
     }
   }, [showModal]);
 
-  const mesh = nodes["Mesh6965_1Ru-exiting"];
+  const mesh = nodes["Mesh2887_Tving-table"];
   const zone = mesh.name.includes("zone") ? mesh.name.slice(-3) : null;
   const meshInfo: IClickedMeshInfo = {
     area_name: areaName,
@@ -78,7 +78,7 @@ export function Exiting1RuModel({
         onClick={() => onMeshClick(meshInfo)}
         onPointerOver={() => onMeshOver(meshInfo)}
         onPointerOut={onMeshOut}
-        position={[-509.209, 25.682, 155.138]}
+        position={[563.362, 139.246, 975.975]}
         rotation={[-3.141, -1.305, -3.141]}
         scale={0.292}
       >
@@ -92,4 +92,4 @@ export function Exiting1RuModel({
   );
 }
 
-useGLTF.preload("/models/exiting-1ru.glb");
+useGLTF.preload("/models/tving-table.glb");
