+++ src/components/elisa/MeshLabel.tsx
@@ -1,7 +1,7 @@
 import { Card } from "../shadcn-ui/card";
-import { TClickedMeshInfo } from "./StadiumModel";
+import { IClickedMeshInfo } from "./3d-models/StadiumModel";
 
-const MeshLabel: React.FC<TClickedMeshInfo> = ({ area_name, zone }) => {
+const MeshLabel: React.FC<IClickedMeshInfo> = ({ area_name, zone }) => {
   return (
     <Card className="rounded-xl border-none p-3 text-lg font-bold bg-white bg-opacity-65 dark:bg-black dark:bg-opacity-55 w-auto min-w-[150px] flex flex-col justify-center items-center">
       <p>{area_name}</p>
