--- src/components/elisa/StadiumSkeleton.tsx
@@ -1,5 +1,0 @@
-const StadiumSkeleton = () => {
-  return <div>StadiumSkeleton</div>;
-};
-
-export default StadiumSkeleton;
