+++ src/utils/supabase/middleware.ts
@@ -2,10 +2,12 @@ import { createServerClient } from '@supabase/ssr'
 import { NextResponse, type NextRequest } from 'next/server'
 
 export async function updateSession(request: NextRequest) {
+  // 기본 응답 객체 생성
   let supabaseResponse = NextResponse.next({
     request,
   })
 
+  // Supabase 서버 클라이언트 생성 (쿠키 기반 인증 유지)
   const supabase = createServerClient(
     process.env.NEXT_PUBLIC_SUPABASE_URL!,
     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
@@ -36,7 +38,7 @@ export async function updateSession(request: NextRequest) {
   const {
     data: { user },
   } = await supabase.auth.getUser() // Refreshing the Auth token
-
+  // 인증되지 않은 사용자는 로그인 페이지로 리디렉트
   if (
     !user &&
     !request.nextUrl.pathname.startsWith('/sign-in') &&
@@ -60,6 +62,6 @@ export async function updateSession(request: NextRequest) {
   //    return myNewResponse
   // If this is not done, you may be causing the browser and server to go out
   // of sync and terminate the user's session prematurely!
-
+  // 최종적으로 갱신된 응답 반환 (쿠키 동기화 유지)
   return supabaseResponse
 }
